+++ sisl/physics/tests/test_hamiltonian.py
@@ -13,7 +13,7 @@ from sisl import Geometry, Atom, SuperCell, Hamiltonian, Spin, BandStructure, Mo
 from sisl import get_distribution
 from sisl import oplist
 from sisl import Grid, SphericalOrbital, SislError
-from sisl.physics.electron import berry_phase, spin_squared, conductivity
+from sisl.physics.electron import berry_phase, spin_contamination, conductivity
 
 
 pytestmark = [pytest.mark.physics, pytest.mark.hamiltonian,
@@ -1355,7 +1355,7 @@ class TestHamiltonian:
         assert np.abs(Hcsr[-1] - Ht.tocsr(-1)).sum() == 0
 
     @pytest.mark.parametrize("k", [[0, 0, 0], [0.1, 0, 0]])
-    def test_spin_squared(self, setup, k):
+    def test_spin_contamination(self, setup, k):
         g = Geometry([[i, 0, 0] for i in range(10)], Atom(6, R=1.01), sc=SuperCell(1, nsc=[3, 1, 1]))
         H = Hamiltonian(g, spin=Spin.POLARIZED)
         H.construct(([0.1, 1.1], [[0, 0.1], [1, 1.1]]))
@@ -1364,39 +1364,39 @@ class TestHamiltonian:
         es_alpha = H.eigenstate(k, spin=0)
         es_beta = H.eigenstate(k, spin=1)
 
-        sup, sdn = spin_squared(es_alpha.state, es_beta.state)
-        sup1, sdn1 = H.spin_squared(k)
+        sup, sdn = spin_contamination(es_alpha.state, es_beta.state)
+        sup1, sdn1 = H.spin_contamination(k)
         assert sup.sum() == pytest.approx(sdn.sum())
         assert np.all(sup1 == sup)
         assert np.all(sdn1 == sdn)
         assert len(sup) == es_alpha.shape[0]
         assert len(sdn) == es_beta.shape[0]
 
-        sup, sdn = spin_squared(es_alpha.sub(range(2)).state, es_beta.state)
+        sup, sdn = spin_contamination(es_alpha.sub(range(2)).state, es_beta.state)
         assert sup.sum() == pytest.approx(sdn.sum())
         assert len(sup) == 2
         assert len(sdn) == es_beta.shape[0]
 
-        sup, sdn = spin_squared(es_alpha.sub(range(3)).state, es_beta.sub(range(2)).state)
-        sup1, sdn1 = H.spin_squared(k, 3, 2)
+        sup, sdn = spin_contamination(es_alpha.sub(range(3)).state, es_beta.sub(range(2)).state)
+        sup1, sdn1 = H.spin_contamination(k, 3, 2)
         assert sup.sum() == pytest.approx(sdn.sum())
         assert np.all(sup1 == sup)
         assert np.all(sdn1 == sdn)
         assert len(sup) == 3
         assert len(sdn) == 2
 
-        sup, sdn = spin_squared(es_alpha.sub(0).state.ravel(), es_beta.sub(range(2)).state)
+        sup, sdn = spin_contamination(es_alpha.sub(0).state.ravel(), es_beta.sub(range(2)).state)
         assert sup.sum() == pytest.approx(sdn.sum())
         assert sup.ndim == 1
         assert len(sup) == 1
         assert len(sdn) == 2
 
-        sup, sdn = spin_squared(es_alpha.sub(0).state.ravel(), es_beta.sub(0).state.ravel())
+        sup, sdn = spin_contamination(es_alpha.sub(0).state.ravel(), es_beta.sub(0).state.ravel())
         assert sup.sum() == pytest.approx(sdn.sum())
         assert sup.ndim == 0
         assert sdn.ndim == 0
 
-        sup, sdn = spin_squared(es_alpha.sub(range(2)).state, es_beta.sub(0).state.ravel())
+        sup, sdn = spin_contamination(es_alpha.sub(range(2)).state, es_beta.sub(0).state.ravel())
         assert sup.sum() == pytest.approx(sdn.sum())
         assert len(sup) == 2
         assert len(sdn) == 1